Update story with proper data
