+++ apps/web/src/components/AssessmentStepTracker/AssessmentStepTracker.test.tsx
@@ -5,166 +5,28 @@ import "@testing-library/jest-dom";
 import React from "react";
 import { Provider as GraphqlProvider } from "urql";
 import { pipe, fromValue, delay } from "wonka";
-import { faker } from "@faker-js/faker";
 import { screen, within } from "@testing-library/react";
 
 import { axeTest, renderWithProviders } from "@gc-digital-talent/jest-helpers";
-import {
-  fakeAssessmentSteps,
-  fakePoolCandidates,
-  fakePools,
-} from "@gc-digital-talent/fake-data";
-import {
-  AssessmentResultType,
-  Pool,
-  PoolSkillType,
-} from "@gc-digital-talent/graphql";
+import { AssessmentDecision } from "@gc-digital-talent/graphql";
 import { notEmpty } from "@gc-digital-talent/helpers";
 
-import {
-  ArmedForcesStatus,
-  AssessmentDecision,
-  AssessmentResult,
-} from "~/api/generated";
+import { NO_DECISION } from "~/utils/assessmentResults";
 
 import AssessmentStepTracker, {
   AssessmentStepTrackerProps,
 } from "./AssessmentStepTracker";
 import { groupPoolCandidatesByStep, sortResultsAndAddOrdinal } from "./utils";
-import { NO_DECISION } from "../../utils/assessmentResults";
-
-const fakePool = fakePools(1)[0];
-const fakePoolAssessmentSteps = fakeAssessmentSteps(2);
-const fakeCandidates = fakePoolCandidates(4);
-
-const getAssessmentResult = (
-  decision?: AssessmentDecision | null,
-): AssessmentResult => ({
-  id: faker.string.uuid(),
-  assessmentDecision:
-    typeof decision === "undefined" ? AssessmentDecision.Successful : decision,
-  assessmentStep: fakePoolAssessmentSteps[0],
-});
-
-const priorityEntitlementCandidate = {
-  ...fakeCandidates[0],
-  id: "priority-entitlement",
-  user: {
-    ...fakeCandidates[0].user,
-    firstName: "priority",
-    lastName: "entitlement",
-    hasPriorityEntitlement: true,
-    armedForcesStatus: ArmedForcesStatus.NonCaf,
-  },
-  isBookmarked: false,
-  assessmentResults: [getAssessmentResult()],
-};
-const armedForcesCandidate = {
-  ...fakeCandidates[1],
-  id: "armed-forces",
-  user: {
-    ...fakeCandidates[1].user,
-    firstName: "armed",
-    lastName: "forces",
-    hasPriorityEntitlement: false,
-    armedForcesStatus: ArmedForcesStatus.Veteran,
-  },
-  isBookmarked: false,
-  assessmentResults: [getAssessmentResult()],
-};
-const bookmarkedCandidate = {
-  ...fakeCandidates[2],
-  id: "bookmarked",
-  user: {
-    ...fakeCandidates[2].user,
-    firstName: "bookmarked",
-    lastName: "candidate",
-    hasPriorityEntitlement: false,
-    armedForcesStatus: ArmedForcesStatus.NonCaf,
-  },
-  isBookmarked: true,
-  assessmentResults: [getAssessmentResult()],
-};
-const unassessedCandidate = {
-  ...fakeCandidates[3],
-  id: "unassessed",
-  user: {
-    ...fakeCandidates[3].user,
-    firstName: "unassessed",
-    lastName: "candidate",
-    hasPriorityEntitlement: false,
-    armedForcesStatus: ArmedForcesStatus.NonCaf,
-  },
-  isBookmarked: false,
-  assessmentResults: [getAssessmentResult(null)],
-};
-const unassessedWithSuccess = {
-  ...unassessedCandidate,
-  assessmentResults: [
-    {
-      ...getAssessmentResult(),
-      type: AssessmentResultType.Skill,
-      poolSkill: {
-        id: faker.string.uuid(),
-        type: PoolSkillType.Nonessential,
-      },
-    },
-  ],
-};
-const lastByFirstName = {
-  ...unassessedWithSuccess,
-  id: "last-by-first-name",
-  isBookmarked: false,
-  user: {
-    id: faker.string.uuid(),
-    firstName: "BB",
-    lastName: "AA",
-    hasPriorityEntitlement: false,
-    armedForcesStatus: ArmedForcesStatus.NonCaf,
-  },
-};
-const firstByName = {
-  ...unassessedWithSuccess,
-  id: "first-by-name",
-  isBookmarked: false,
-  user: {
-    id: faker.string.uuid(),
-    firstName: "AA",
-    hasPriorityEntitlement: false,
-    armedForcesStatus: ArmedForcesStatus.NonCaf,
-  },
-};
-
-const testCandidates = [
-  priorityEntitlementCandidate,
+import {
   armedForcesCandidate,
   bookmarkedCandidate,
-  unassessedCandidate,
-  unassessedWithSuccess,
-  lastByFirstName,
   firstByName,
-];
-
-// eslint-disable-next-line import/prefer-default-export
-export const poolWithAssessmentSteps: Pool = {
-  ...fakePool,
-  assessmentSteps: [
-    {
-      ...fakePoolAssessmentSteps[1],
-      sortOrder: 2,
-      assessmentResults: [],
-    },
-    {
-      ...fakePoolAssessmentSteps[0],
-      sortOrder: 1,
-      assessmentResults: testCandidates.map((candidate) => ({
-        ...candidate.assessmentResults[0],
-        poolCandidate: candidate,
-      })),
-    },
-  ],
-  poolCandidates: testCandidates,
-};
+  lastByFirstName,
+  poolWithAssessmentSteps,
+  priorityEntitlementCandidate,
+  testCandidates,
+  unassessedCandidate,
+} from "./testData";
 
 // This should always make the component visible
 const defaultProps: AssessmentStepTrackerProps = {
